+++ jpos/src/main/java/org/jpos/transaction/participant/BSHTransactionParticipant.java
@@ -106,7 +106,7 @@ public class BSHTransactionParticipant extends SimpleLogSource
         int result = ABORTED | READONLY;
         if (prepareMethod != null) {
             try {
-                result = ((Integer) executeMethod(prepareMethod, id, context, ev, "result")).intValue();
+                result = (Integer) executeMethod(prepareMethod, id, context, ev, "result");
             } catch (Exception ex) {
                 ev.addMessage(ex);
             }
@@ -124,7 +124,7 @@ public class BSHTransactionParticipant extends SimpleLogSource
         int result = ABORTED | READONLY;
         if (prepareForAbortMethod != null) {
             try {
-                result = ((Integer) executeMethod(prepareForAbortMethod, id, context, ev, "result")).intValue();
+                result = (Integer) executeMethod(prepareForAbortMethod, id, context, ev, "result");
             } catch (Exception ex) {
                 ev.addMessage(ex);
             }
@@ -155,7 +155,7 @@ public class BSHTransactionParticipant extends SimpleLogSource
     throws EvalError, FileNotFoundException, IOException {
         Map params = new HashMap();
         params.put("context", context);
-        params.put("id", Long.valueOf(id));
+        params.put("id", id);
         params.put("evt", evt);
         params.put("self", this);
         return m.execute(params, resultName);
